Python App/requirements.txt
Flask==3.1.1
python-dotenv==1.1.0
Flask-SQLAlchemy==3.1.1
Flask-PyMongo==3.0.1
pymongo==4.13.2
pyodbc==5.2.0
pandas==2.2.2
numpy==1.26.4